# openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Article - OpenAPI 3.0
  description: |-
    This is an API for Article application.
  version: 1.0.1
...
paths:
  /user:
    post:
      tags:
        - user
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createUser
      requestBody:
        $ref: '#/components/requestBodies/UserCreate'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Not valid input data 
  /user/login:
    get:
      tags:
        - user
      summary: Logs user into the system
      description: ''
      operationId: loginUser
      parameters:
        - name: username
          in: query
          description: The user name for login
          required: false
          schema:
            type: string
        - name: password
          in: query
          description: The password for login in clear text
          required: false
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                type: string
        '400':
          description: Invalid username/password supplied
...
  /article:
    post:
      tags:
        - article
      summary: Add a new article to the site
      description: Add a new article to the site
      operationId: addArticle
      requestBody:
        $ref: '#/components/requestBodies/ArticleCreate'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
            application/xml:
              schema:
                $ref: '#/components/schemas/Article'
        '403':
          description: Not enough permissions
        '400':
          description: Invalid input
      security:
        - article_auth:
            - admin
            - moderator
...
  /change:
    post:
      tags:
        - change
      summary: Add a new change to the site
      description: Add a new change to the site
      operationId: addChange
      requestBody:
        $ref: '#/components/requestBodies/ChangeCreate'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Change'
            application/xml:
              schema:
                $ref: '#/components/schemas/Change'
        '403':
          description: Not enough permissions
        '404':
          description: Article not found
        '400':
          description: Invalid input
      security:
        - article_auth:
            - admin
            - moderator
            - user
...
  /mychanges:
    get:
      tags:
        - change
      summary: Find all changes done by current user.
      description: Returns array of changes
      operationId: getUserChanges
      responses:
        '200':
          description: successful operation
          content:
            application\json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Change'
        '403':
          description: Unathorized
        '404':
          description: Changes not found
      security:
        - article_auth:
            - admin
            - moderator
            - user
  /changesInReview:
    get:
      tags:
        - change
      summary: Find all changes that need a review. 
      description: Returns array of changes
      operationId: getChangesInReview
      responses:
        '200':
          description: successful operation
          content:
            application\json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Change'
        '403':
          description: Unathorized
        '404':
          description: Changes not found
      security:
        - article_auth:
            - admin
            - moderator
  /review:
    post:
      tags:
        - review
      summary: Create review  
      description: This can only be done by moderator or admin. If verdict 1, change article text to newText of the change, increment version of the article, set change status accepted and set status of all changes with the same articleId and version to denied. Otherwise only set change status denied.
      operationId: createReview
      requestBody:
        $ref: '#/components/requestBodies/ReviewCreate'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
            application/xml:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          description: Change not found
        '403':
          description: Unathorized
      security:
        - article_auth:
          - admin
          - moderator
...
  /myReviews:
    get:
      tags:
        - review
      summary: Get list of reviews done by user  
      description: Returns list of reviews done by user
      operationId: listReviewsDone
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/Review'
        '404':
          description: Reviews not found
        '403':
          description: Not enough permissions
      security:
        - article_auth:
          - admin
          - moderator
  /myChangesReviewed:
    get:
      tags:
        - review
      summary: Get list of reviews of changes done by user  
      description: Returns list of reviews of changes done by user
      operationId: listReviewsByChangesDone
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/Review'
        '404':
          description: Reviews not found
        '403':
          description: Unathorized
      security:
        - article_auth:
          - admin
          - moderator
          - user
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        username:
          type: string
          example: theUser
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: James
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: '12345'
        phone:
          type: string
          example: '12345'
        userStatus:
          type: integer
          description: User Status
          format: int32
          example: 1
      xml:
        name: user
    Article:
      required:
        - id
        - name
        - text
        - version
        - creatorId
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: Harry Potter
        text:
                type: string
                description: text of an arcticle
                example: "Harry Potter is a series of seven fantasy novels written by British author J. K. Rowling. The novels chronicle the lives of a young wizard, Harry Potter, and his friends Hermione Granger and Ron Weasley, all of whom are students at Hogwarts School of Witchcraft and Wizardry. The main story arc concerns Harry's conflict with Lord Voldemort, a dark wizard who intends to become immortal, overthrow the wizard governing body known as the Ministry of Magic and subjugate all wizards and Muggles (non-magical people)."
        version:
          type: integer
          format: int64
          example: 0
        creatorId:
          type: integer
          format: int64
          example: 4
      xml:
        name: article
    Change:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        articleId:
          type: integer
          format: int64
          example: 25
        articleVersion:
          type: integer
          format: int64
          example: 3
        oldText:
          type: string
          example: "Harry Potter is a series of seven fantasy novels written by British author J. K. Rowling. The novels chronicle the lives of a young wizard, Harry Potter, and his friends Hermione Granger and Ron Weasley, all of whom are students at Hogwarts School of Witchcraft and Wizardry. The main story arc concerns Harry's conflict with Lord Voldemort, a dark wizard who intends to become immortal, overthrow the wizard governing body known as the Ministry of Magic and subjugate all wizards and Muggles (non-magical people)."
        newText:
          type: string
          example: "Harry Potter is a series of seven fantasy novels written by British author J. K. Rowling. The novels chronicle the lives of a young wizard, Harry Potter, and his friends Hermione Granger and Ron Weasley, all of whom are students at Hogwarts School of Witchcraft and Wizardry. The main story arc concerns Harry's conflict with Lord Voldemort, a dark wizard who intends to become immortal, overthrow the wizard governing body known as the Ministry of Magic and subjugate all wizards and Muggles (non-magical people). It is my favorite series."
        status:
          type: string
          example: InReview
        proposerId:
          type: integer
          format: int64
          example: 2
      xml:
        name: change
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        changeId:
          type: integer
          format: int64
          example: 59
        verdict:
          type: integer
          format: int64
          example: 0
        comment:
          type: string
          example: Can`t accept this changes, because it has too many mistakes.
        reviewerId:
          type: integer
          format: int64
          example: 2
      xml:
        name: review  
...
  requestBodies:
    ChangeCreate:
      description: Create change object
      required: true
      content:
        application/json:
            schema:
              type: object
              properties:
                articleId:
                  type: integer
                  format: int64
                  example: 25
                newText:
                  type: string
                  example: "Harry Potter is a series of seven fantasy novels written by British author J. K. Rowling. The novels chronicle the lives of a young wizard, Harry Potter, and his friends Hermione Granger and Ron Weasley, all of whom are students at Hogwarts School of Witchcraft and Wizardry. The main story arc concerns Harry's conflict with Lord Voldemort, a dark wizard who intends to become immortal, overthrow the wizard governing body known as the Ministry of Magic and subjugate all wizards and Muggles (non-magical people). It is my favorite series."
              xml:
                name: change
    UserCreate:
      description: Create user object
      required: true
      content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: theUser
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: James
                email:
                  type: string
                  example: john@email.com
                password:
                  type: string
                  example: '12345'
                phone:
                  type: string
                  example: '12345'
              xml:
                name: user
    ArticleCreate:
      description: Create a new article in the site
      required: true
      content:
        application/json:
          schema:
            required:
              - name
              - text
            type: object
            properties:
              name:
                type: string
                example: Harry Potter
              text:
                type: string
                description: text of an arcticle
                example: "Harry Potter is a series of seven fantasy novels written by British author J. K. Rowling. The novels chronicle the lives of a young wizard, Harry Potter, and his friends Hermione Granger and Ron Weasley, all of whom are students at Hogwarts School of Witchcraft and Wizardry. The main story arc concerns Harry's conflict with Lord Voldemort, a dark wizard who intends to become immortal, overthrow the wizard governing body known as the Ministry of Magic and subjugate all wizards and Muggles (non-magical people)."
            xml:
              name: article
    ReviewCreate:
      description: Create review object
      required: true
      content:
          application/json:
            schema:
              type: object
              properties:
                changeId:
                  type: integer
                  format: int64
                  example: 59
                verdict:
                  type: integer
                  format: int64
                  example: 0
                comment:
                  type: string
                  example: Can`t accept this changes, because it has too many mistakes.
              xml:
                name: review
  securitySchemes:
    article_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: url
          scopes:
            admin : can do all actions
            moderator : can do reviews
            user : can see articles, change own profile and create changes in articles 